+++ warp/native/builtin.h
@@ -1384,26 +1384,18 @@ inline CUDA_CALLABLE T atomic_add(T* buf, T value)
 #endif
 }
 
-// emulate atomic int64 add with atomicCAS()
 template <>
-inline CUDA_CALLABLE int64 atomic_add(int64* address, int64 val)
+inline CUDA_CALLABLE int64 atomic_add(int64* buf, int64 value)
 {
-#if defined(__CUDA_ARCH__)
-    unsigned long long int *address_as_ull = (unsigned long long int*)address;
-    unsigned long long int old = *address_as_ull, assumed;
-
-    while (val < (int64)old)
-    {
-        assumed = old;
-        old = atomicCAS(address_as_ull, assumed, (int64)val);
-    }
-
-    return (int64)old;
-
-#else
-    int64 old = *address;
-    *address = min(old, val);
+#if !defined(__CUDA_ARCH__)
+    int64 old = buf[0];
+    buf[0] += value;
     return old;
+#else  // CUDA compiled by NVRTC
+    unsigned long long int *buf_as_ull = (unsigned long long int*)buf;
+    unsigned long long int unsigned_value = static_cast<unsigned long long int>(value);
+    unsigned long long int result = atomicAdd(buf_as_ull, unsigned_value);
+    return static_cast<int64>(result);
 #endif
 }