+++ CHANGELOG.md
@@ -73,6 +73,7 @@
 - Update built-in documentation to accurately reflect their differentiability status
   ([GH-970](https://github.com/NVIDIA/warp/issues/970)).
 - Fix handling of multi-line lambda expressions and lambda expressions involving parentheses in `wp.map()` ([GH-984](https://github.com/NVIDIA/warp/issues/984)).
+- Fix `wp.atomic_add()` for int64 type ([GH-977](https://github.com/NVIDIA/warp/issues/977))
 
 ## [1.9.0] - 2025-09-04